+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # CHANGELOG
 
+## v1.2.1, 2025-02-27
+
+- New standard textbook option for logLei discontinuity fix
+- Support for changing BDE integrator consolidation interval
+- Support for ignoring Jacobians in the diffusion stencil
+- Begun refactor of CustomModel (now renamed Model, with CustomModel alias still available)
+
+### Breaking Changes
+
+- N/A
+
+### Deprecations
+
+- N/A
+
+### New Features
+
+- New standard textbook option to remove the discontinuity in the NRL formulary version of e-i Coulomb Log
+- The BDE integrator consolidation interval can now be set
+- ignoreJacobian option in diffusion stencil
+
+### Bug Fixes
+
+- e-i collision operator ion density dependence fixed
+
+### New Notebooks 
+
+- N/A
+
 ## v1.2.0, 2024-09-19
 
 - Added CVODE integrator as an option